requirements.txt
asgiref==3.11.0
astroid==4.0.2
boto3==1.42.11
botocore==1.42.11
dill==0.4.0
Django==5.2.9
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
isort==7.0.0
jmespath==1.0.1
mccabe==0.7.0
packaging==25.0
platformdirs==4.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
pylint==4.0.4
python-dateutil==2.9.0.post0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
tomlkit==0.13.3
urllib3==2.6.2
